requirements.txt
amqp
asgiref
async-timeout
attrs
autobahn
Automat
billiard
celery
cffi
channels
channels-redis
click
click-didyoumean
click-plugins
click-repl
colorama
constantly
cryptography
daphne
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
et-xmlfile
gevent
greenlet
gunicorn
hyperlink
idna
incremental
inflection
jsonschema
jsonschema-specifications
kombu
Markdown
msgpack
numpy
openpyxl
packaging
pandas
prompt-toolkit
psycopg2
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
python-dateutil
python-dotenv
pytz
PyYAML
redis
referencing
rpds-py
service-identity
six
sqlparse
swagger-spec-validator
Twisted
txaio
typing_extensions
tzdata
uritemplate
vine
wcwidth
XlsxWriter
